State Machine working!
Added a finite state machine. I think it's a bit weird that I only have to use 1 method in abstract class PlayerState.cs. called ChangeState(). Not sure if I implemented the state machine correctly but it is working and adding new states shouldn't be that hard. I have to call the ChangeState() method that is located in the PlayerState class in the PlayerController.cs Update method for this state machine to work. I hope this is ok.

If there are multiple drinks player can now drink one of them without the other drinks current volume variable change.